requirements.txt
pandas==2.2.3
requests==2.32.3
lxml==5.3.0
pymongo==4.10.1
python-dotenv==1.0.0
